+++ WHATS_NEW.md
@@ -1,4 +1,8 @@
-### Added
+### Added 
 
-*  room_message_updated
+ * room_message_updated
+
+### Removed
+
+*  custom message INTENT_DETECTED